[FE/#74] Refactor : edit handler 재사용 위해 분리
